requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
openai==1.58.1
langchain==0.3.13
langchain-openai==0.3.0
langchain-text-splitters==0.3.3
langgraph==0.2.60
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
alembic==1.14.0
httpx==0.28.1
python-multipart==0.0.20
pyjwt==2.9.0
bcrypt==4.2.1
pinecone-client==5.0.1
tiktoken==0.7.0
pymupdf==1.27.2
python-docx==1.2.0
python-pptx==1.0.2
openpyxl==3.1.5
